+++ src/sharp.cc
@@ -23,6 +23,7 @@
 
 static void* sharp_vips_init(void*) {
   g_setenv("VIPS_MIN_STACK_SIZE", "2m", FALSE);
+  g_setenv("PANGOCAIRO_BACKEND", "fontconfig", FALSE);
   vips_init("sharp");
   return nullptr;
 }